server/requirements.txt

# CMAKE_ARGS="-DLLAMA_METAL=on" pip install llama-cpp-python
llama-cpp-python
